backend/requirements.txt
# Flask Backend Dependencies
# Install with: pip install -r requirements.txt

# Core Flask
Flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.10.0
python-engineio==4.8.1

# MongoDB
flask-pymongo==2.3.0
pymongo==4.6.1

# Authentication
flask-jwt-extended==4.6.0
bcrypt==4.1.2

# NLP & AI - Local Offline Models
transformers==4.36.2
torch==2.1.2 --index-url https://download.pytorch.org/whl/cpu
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4

# PDF & Document Processing
pypdf==6.7.3
pdfminer.six==20231228
pdfplumber==0.11.9
PyMuPDF==1.24.1
python-docx==1.1.0
pdfkit==1.0.0
reportlab==4.0.7

# Utilities
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
eventlet==0.34.2

# Development
pytest==7.4.3
black==23.12.1